Started experimenting with Different Arrangements
As a way or approaching some different layouts, I asked copilot to create a new js file (arrangements.js) and start by offering an animated shuffle. While putting the images in a random order isn't very useful it allows me to see  a "whole document" animation.  The code doesn't actually work. The key function is called (upon space bar) but nothing happens, though if I resize the window and it redraws, everything is immediately in its shuffled position. So, the animation isn't happening.